YUK-12824 Add Demand Reset collection action supporting both RF and PLC devices/ backing service

+++ yukon-client/common/src/com/cannontech/common/bulk/collection/DeviceIdListCollectionProducer.java
@@ -112,7 +112,7 @@ public class DeviceIdListCollectionProducer implements DeviceCollectionProducer 
             }
 
             @Override
-            public long getDeviceCount() {
+            public int getDeviceCount() {
                 return deviceIds.size();
             }
